+++ processor/src/main/java/ca/bc/gov/app/entity/BaseEntity.java
@@ -1,4 +1,4 @@
-package ca.bc.gov.app.entity.client;
+package ca.bc.gov.app.entity;
 
 import java.time.LocalDateTime;
 import lombok.AllArgsConstructor;
