+++ processor/src/main/java/ca/bc/gov/app/entity/ExpirableBaseEntity.java
@@ -1,4 +1,4 @@
-package ca.bc.gov.app.entity.client;
+package ca.bc.gov.app.entity;
 
 import jakarta.validation.constraints.NotNull;
 import java.time.LocalDate;
