+++ processor/src/main/java/ca/bc/gov/app/entity/SubmissionLocationContactIdEntity.java
@@ -1,21 +1,19 @@
-package ca.bc.gov.app.entity.client;
+package ca.bc.gov.app.entity;
 
-import ca.bc.gov.app.ApplicationConstant;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 import lombok.With;
 import org.springframework.data.relational.core.mapping.Column;
-import org.springframework.data.relational.core.mapping.Table;
 
-@Table(name = "submission_location_contact_xref", schema = ApplicationConstant.POSTGRES_ATTRIBUTE_SCHEMA)
+
 @Data
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
 @With
-public class SubmissionLocationContactEntity {
+public class SubmissionLocationContactIdEntity {
 
   @Column("submission_contact_id")
   private Integer submissionContactId;
