+++ processor/src/main/java/ca/bc/gov/app/repository/CountryCodeRepository.java
@@ -1,6 +1,6 @@
-package ca.bc.gov.app.repository.client;
+package ca.bc.gov.app.repository;
 
-import ca.bc.gov.app.entity.client.CountryCodeEntity;
+import ca.bc.gov.app.entity.CountryCodeEntity;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.repository.reactive.ReactiveCrudRepository;
 import org.springframework.data.repository.reactive.ReactiveSortingRepository;
