+++ processor/src/main/java/ca/bc/gov/app/repository/SubmissionLocationContactRepository.java
@@ -1,7 +1,7 @@
-package ca.bc.gov.app.repository.client;
+package ca.bc.gov.app.repository;
 
-import ca.bc.gov.app.entity.client.SubmissionLocationContactEntity;
-import ca.bc.gov.app.entity.client.SubmissionLocationContactIdEntity;
+import ca.bc.gov.app.entity.SubmissionLocationContactEntity;
+import ca.bc.gov.app.entity.SubmissionLocationContactIdEntity;
 import org.springframework.data.repository.reactive.ReactiveCrudRepository;
 import org.springframework.stereotype.Repository;
 import reactor.core.publisher.Flux;
