+++ processor/src/main/java/ca/bc/gov/app/repository/SubmissionMatchDetailRepository.java
@@ -1,6 +1,6 @@
-package ca.bc.gov.app.repository.client;
+package ca.bc.gov.app.repository;
 
-import ca.bc.gov.app.entity.client.SubmissionMatchDetailEntity;
+import ca.bc.gov.app.entity.SubmissionMatchDetailEntity;
 import org.springframework.data.repository.reactive.ReactiveCrudRepository;
 import org.springframework.stereotype.Repository;
 import reactor.core.publisher.Flux;
